refactor: use f64 for intermediate computation to fix precision drift
Switches coverage buffers, aggregation, MatrixRow storage, and sort
metrics from f32 to f64.  BigWig values (on-disk f32) are promoted to
f64 on read; formatting accepts f64 directly, removing the redundant
as-f64 cast.  Fixes gtf_input parity test (max delta 7e-6 → <1e-10).

+++ src/pipeline/core/collector.rs
@@ -127,7 +127,7 @@ mod tests {
     use crate::pipeline::matrix::MatrixHeader;
     use std::sync::Arc;
 
-    fn dummy_row(name: &str, values: Vec<f32>) -> MatrixRow {
+    fn dummy_row(name: &str, values: Vec<f64>) -> MatrixRow {
         MatrixRow {
             record: BedRecord {
                 chrom: Arc::from("chr1"),
