+++ src/pipeline/core/traits.rs
@@ -68,7 +68,7 @@ pub trait PipelineMode: Sync {
     fn postprocess_row(
         &self,
         record: BedRecord,
-        values: Vec<f32>,
+        values: Vec<f64>,
         sample_count: usize,
         bin_count: usize,
         metadata: &Self::Metadata,
