+++ src/pipeline/reference_point.rs
@@ -90,7 +90,7 @@ impl PipelineMode for ReferencePointMode {
     fn postprocess_row(
         &self,
         record: BedRecord,
-        values: Vec<f32>,
+        values: Vec<f64>,
         sample_count: usize,
         bin_count: usize,
         _metadata: &Self::Metadata,
